+++ lib/cgo/api.client.go
@@ -726,3 +726,85 @@ func SKY_api_Client_DecryptWallet(_c C.Client__Handle, _id string, _password str
 	}
 	return
 }
+
+//export SKY_api_Client_CreateWallet
+func SKY_api_Client_CreateWallet(_c C.Client__Handle, _cwo C.CreateWalletOptions__Handle, _arg2 *C.WalletResponse__Handle) (____error_code uint32) {
+	c, okc := lookupClientHandle(_c)
+	if !okc {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+	cwo, okcwo := lookupCreateWalletOptionsHandle(_cwo)
+	if !okcwo {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+	wr, ____return_err := c.CreateWallet(*cwo)
+	____error_code = libErrorCode(____return_err)
+	if ____return_err == nil {
+		*_arg2 = registerWalletResponseHandle(wr)
+	}
+	return
+}
+
+//export SKY_api_Client_InjectTransactionNoBroadcast
+func SKY_api_Client_InjectTransactionNoBroadcast(_c C.Client__Handle, _txn C.Transaction__Handle, _arg0 *C.GoString_) (____error_code uint32) {
+	c, okc := lookupClientHandle(_c)
+	if !okc {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+	txn, oktxn := lookupTransactionHandle(_txn)
+	if !oktxn {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+	arg0, ____return_err := c.InjectTransactionNoBroadcast(txn)
+	____error_code = libErrorCode(____return_err)
+	if ____return_err == nil {
+
+		copyString(arg0, _arg0)
+	}
+	return
+}
+
+//export SKY_api_Client_InjectEncodedTransactionNoBroadcast
+func SKY_api_Client_InjectEncodedTransactionNoBroadcast(_c C.Client__Handle, rawtx string, _arg0 *C.GoString_) (____error_code uint32) {
+	c, okc := lookupClientHandle(_c)
+	if !okc {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+
+	arg0, ____return_err := c.InjectEncodedTransactionNoBroadcast(rawtx)
+	____error_code = libErrorCode(____return_err)
+	if ____return_err == nil {
+
+		copyString(arg0, _arg0)
+	}
+	return
+}
+
+//export SKY_api_Client_RecoverWallet
+func SKY_api_Client_RecoverWallet(_c C.Client__Handle, _req C.WalletRecoverRequest__Handle, _arg2 *C.WalletResponse__Handle) (____error_code uint32) {
+	c, okc := lookupClientHandle(_c)
+	if !okc {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+
+	req, okreq := lookupWalletRecoverRequestHandle(_req)
+	if !okreq {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+
+	arg2, ____return_err := c.RecoverWallet(*req)
+	____error_code = libErrorCode(____return_err)
+
+	if ____return_err == nil {
+		*_arg2 = registerWalletResponseHandle(arg2)
+	}
+	return
+
+}